Associate the exclude_rt with the correct interface.
Every exclude_rt was derived from the system's default route (which is
retrieved using ip route show default). This causes the
communication to use the default gateway even if the system had a more
preferred route for the destination. In the worst case, the destination
is unreachable from the default route, causing loss of communications.

This fix addresses this behavior by looking up the system's preferred
route for the destination and then deriving the exclude_rt from that
route.

A secondary benefit of this approach is that the ziti-edge-tunnel can use
indirect underlays such as Speedify or Tailscale to communicate with the
controllers and edge routers. One of the benefits of indirect underlays
is increasing the likelihood communications pass firewalls and hard
NATs.

Signed-off-by: Tom Carroll <4632752+tomc797@users.noreply.github.com>