Disregard route if update is a subnet of tun's address.
As the tun_ip is derived from dns_block, set the tun's address prefix
length equal to dns_block prefix length. Then ignore any route updates
that are a subnet of the tun's address. This becomes helpful when large
number of DNS intercepts are employed (say wildcard), each causing a /32
route to be added.

Signed-off-by: Tom Carroll <4632752+tomc797@users.noreply.github.com>