Manage routes in a routing table dedicated to ZET.
ZET routes are managed in a dedicated routing table, distinct from main and default. The routing policy database is configured to prefer the routes managed in the ZET table over those stored in the main and default routing tables. This approach provides isolation, and thus additional security for ZET routes.

Signed-off-by: Tom Carroll <4632752+tomc797@users.noreply.github.com>